[add-precommit-autofix-ci] refactor: split dev docker commands by responsibility
